Refactor `BillsCreateScreen` to use the `BillsForm` component
