+++ lib/ui/bills/create/bills_create_screen.dart
@@ -1,189 +1,34 @@
+import 'package:bills_reminder/domain/models/bill.dart';
 import 'package:bills_reminder/ui/bills/create/bills_create_view_model.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:intl/intl.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
+import 'package:bills_reminder/ui/core/bills/bills_form.dart';
 
-class BillsCreateScreen extends StatefulWidget {
+class BillsCreateScreen extends StatelessWidget {
   const BillsCreateScreen({super.key});
 
   @override
-  State<BillsCreateScreen> createState() => _BillsCreateScreenState();
-}
-
-class _BillsCreateScreenState extends State<BillsCreateScreen> {
-  final _formKey = GlobalKey<FormState>();
-  final _nameController = TextEditingController();
-  final _valueController = TextEditingController();
-  final _dateController = TextEditingController();
-
-  DateTime _selectedDate = DateTime.now();
-  bool _notificationEnabled = false;
-  bool _recurrenceEnabled = false;
-  late BillsCreateViewModel _viewModel;
-
-  @override
-  void initState() {
-    super.initState();
-    _dateController.text = DateFormat.yMMMd().format(_selectedDate);
-    _viewModel = BillsCreateViewModel(repository: context.read());
-  }
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _valueController.dispose();
-    _dateController.dispose();
-
-    super.dispose();
-  }
-
-  Future<void> _selectDate(BuildContext context) async {
-    final DateTime? picked = await showDatePicker(
-      context: context,
-      initialDate: _selectedDate,
-      firstDate: DateTime.now(),
-      lastDate: DateTime(DateTime.now().year + 1),
-    );
-
-    if (picked != null && picked != _selectedDate) {
-      setState(() {
-        _selectedDate = picked;
-        _dateController.text = DateFormat.yMMMd().format(_selectedDate);
-      });
-    }
-  }
-
-  @override
   Widget build(BuildContext context) {
+    final viewModel = BillsCreateViewModel(repository: context.read());
+
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Create Bill'),
-        actions: [
-          IconButton(
-            onPressed: () {
-              if (_formKey.currentState!.validate()) {
-                _viewModel.createBill(
-                  name: _nameController.text,
-                  value: double.parse(_valueController.text),
-                  date: _selectedDate,
-                  notification: _notificationEnabled,
-                  recurrence: _recurrenceEnabled,
-                );
+      appBar: AppBar(title: const Text('Create Bill')),
+      body: BillsForm(
+        onSave: (Bill bill) async {
+          await viewModel.createBill(
+            name: bill.name,
+            value: bill.value,
+            date: bill.date,
+            notification: bill.notification,
+            recurrence: bill.recurrence,
+          );
 
-                if (context.mounted) {
-                  context.pop();
-                }
-              }
-            },
-            icon: const Icon(Icons.check),
-            tooltip: 'Create Bill',
-          ),
-        ],
-      ),
-      body: Form(
-        key: _formKey,
-        child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Column(
-              children: <Widget>[
-                TextFormField(
-                  autofocus: true,
-                  controller: _nameController,
-                  decoration: const InputDecoration(
-                    labelText: 'Bill Name',
-                    border: OutlineInputBorder(),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter a bill name';
-                    }
-                    return null;
-                  },
-                  textInputAction: TextInputAction.next,
-                ),
-                const SizedBox(height: 16),
-                TextFormField(
-                  controller: _valueController,
-                  keyboardType: TextInputType.number,
-                  decoration: const InputDecoration(
-                    labelText: 'Bill Value',
-                    border: OutlineInputBorder(),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter a bill value';
-                    }
-                    if (double.tryParse(value) == null) {
-                      return 'Please enter a valid number';
-                    }
-                    return null;
-                  },
-                  textInputAction: TextInputAction.next,
-                  inputFormatters: [
-                    FilteringTextInputFormatter.allow(
-                      RegExp(r'^\d+\.?\d{0,2}'),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 16),
-                TextFormField(
-                  controller: _dateController,
-                  decoration: const InputDecoration(
-                    labelText: 'Date',
-                    border: OutlineInputBorder(),
-                    suffixIcon: Icon(Icons.calendar_today),
-                  ),
-                  readOnly: true,
-                  onTap: () => _selectDate(context),
-                ),
-                const SizedBox(height: 16),
-                CheckboxListTile(
-                  title: const Text('Notification'),
-                  value: _notificationEnabled,
-                  onChanged: (value) {
-                    setState(() {
-                      _notificationEnabled = value!;
-                    });
-                  },
-                ),
-                CheckboxListTile(
-                  title: const Text('Monthly Recurrence'),
-                  value: _recurrenceEnabled,
-                  onChanged: (value) {
-                    setState(() {
-                      _recurrenceEnabled = value!;
-                    });
-                  },
-                ),
-                const SizedBox(height: 16),
-                SizedBox(
-                  width: double.infinity,
-                  child: ElevatedButton(
-                    onPressed: () async {
-                      if (_formKey.currentState!.validate()) {
-                        await _viewModel.createBill(
-                          name: _nameController.text,
-                          value: double.parse(_valueController.text),
-                          date: _selectedDate,
-                          notification: _notificationEnabled,
-                          recurrence: _recurrenceEnabled,
-                        );
-
-                        if (context.mounted) {
-                          context.pop();
-                        }
-                      }
-                    },
-                    child: const Text('Create Bill'),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
+          if (context.mounted) {
+            context.pop();
+          }
+        },
+        isEdit: false,
       ),
     );
   }
